refactor(server): split index.ts into route modules
index.ts was 541 lines of route handlers, session management, and
SPA wiring. It is now 42 lines of wiring: middleware, router mounts,
static serving, SPA fallback, listen.

Route handlers move to server/routes/{game,user,daily,leaderboard}.ts,
each exporting an express.Router. Session state (the in-memory Map,
cleanup interval, middleware) moves to server/session.ts. Daily
constants and date helpers move to services/dailyConfig.ts so both
the daily and leaderboard routers can share them without pulling in
each other. Smoke-tested endpoints return the same shapes at the
same paths.

server/CLAUDE.md gets a routes rule (handlers live in routes/;
index.ts is wiring-only; mount the more specific prefix first).
